support trailers in http1 chunked backend responses
== Usage / VCL ==

Trailer processing is enabled with ``set bereq.http.TE = "trailers";``
in ``vcl_backend_fetch{}``. Only then may the backend send a chunked
response with trailers. Any ``TE`` client header remains filtered, so
trailer processing is neither implicitly enabled nor would trailers
ever be sent downstream (except in pipe mode).

The above only allows the backend to send trailers, whether or not it
actually makes use of that capability (or if it even chooses chunked
encoding) is up to the backend.

For processing of the backend response, we enable trailer handling
only if a ``Trailers:`` backend response header is present - either
from the backend or set in ``vcl_backend_response{}``. ``Trailers:``
must contain header names to be accepted in the chunked trailer. In
addition, ``set beresp.http.Trailers = "*";`` can be used in vcl to
instruct varnish to accept any header in the chunked
trailer. ``Trailers: *`` is never accepted from a backend.

Any accepted trailer headers are copied into the object without
additional VCL control. In other words, any headers allowed by the
``Trailers`` header are accepted as-are. Additional filtering is
possible on the client side.

Trailer processing implicitly disables streaming. The backend response
is read into cache and trailer headers are delivered as normal
response headers.

== Implementation notes ==

The object API requires that the size of the attribute space is known
at object creation time, so we need to guess the amount of space to
reserve for trailers. For this initial implementation, the reserve is
hardcoded as 1KB and we'll probably want some VCL control later.

For trailer processing, the setting OA_HEADERS now needs happen after
reading the body, obviously. It is unclear if The Proprietary
Stevedore (tm) supports this.

To avoid code duplication for header dissection, http1_dissect_hdrs()
is now generalised and declared in cache.h as HTTP1_DissectHdrs()

For logging new/deleted headers from vbf, http_VSLH() and
http_VSLH_del() are now declared in cache_http.h