generalize the worker pool reserve to avoid deadlocks
Previously, we used a minimum number of idle threads (the reserve) to
ensure that we do not assign all threads with client requests and no threads
left over for backend requests.

This was actually only a special case of the more general issue exposed
by h2: Lower priority tasks depend on higher priority tasks (for h2,
sessions need streams, which need requests, which may need backend
requests).

To solve this problem, we divide the reserve by the number of priority
classes and schedule lower priority tasks only if there are enough idle
threads to run higher priority tasks eventually.

This change does not guarantee any upper limit on the amount of time it
can take for a task to be scheduled (e.g. backend requests could be
blocking on arbitrarily long timeouts), so the thread pool watchdog is
still warranted. But this change should guarantee that we do make
progress eventually.

Fixes #2418 for real