Save writev() calls for high numbers of http headers
Since 7bf932f3a674fa934c1bfd586686686003eec550 we might exceed the
number of iovs for many http headers (>31 with default parameters), so
we would issue more than one writev() for the http headers.

While sending http headers, we can use all of the thread workspace for
IOVs, so if we know that we would need a seond writev call anyway, we
re-open V1L to ensure we send as many headers at once as possible.