VTP: Add clone function and per-connection preamble
* preamble

The use case is to open connections with a PROXY header prefix, for
example to open https connections through haproxy.

The preamble is an arbitrary byte stream sent once when the connection
is opened.

It is considered an attribute of the pool, so it gets passed as
arguments to VTP_Ref(). As with the other pool arguments, it is the
caller's responsibility to ensure that the ``id`` argument be unique
also with respect to the preamble.

Compared to other options on how to implement a preamble at thigher
levels, the implementation at the pool level simplifies caller code
by avoidance of duplication and not having to special case new vs.
recycled connections from the pool.

* VTP_Clone()

The use case is to create a new pool based on an existing one, but with
a different id and preamble. This is motivated by the use case of the
preamble, where a single endpoint may be used for many connections with
different preambles.

The implementation is motivated by the fact that opaque handling of the
address attributes of the pool (ip4, ip6, uds) simplifies caller code.