vcc_acl_merge parameter: remove subnets and merge adjacent entries
function:

With the vcc_acl_merge parameter enabled (default: disabled, keeping
the existing behavior), ACLs are optimized in that subnets contained
in other entries are skipped (e.g.  if 1.2.3.0/24 is part of the ACL,
an entry for 1.2.3.128/25 will not be added) and adjacent entries get
merged (e.g.  if both 1.2.3.0/25 and 1.2.3.128/25 are added, they will
be merged to 1.2.3.0/24).

Skip and merge operations on VCL entries are output as warnings during
VCL compilation as entries from the VCL are processed in order.

Logging under the VCL_acl tag can change with this parameter enabled:
Matches on skipped subnet entries are now logged as matches on the
respective supernet entry. Matches on merged entries are logged with a
shorter netmask which might not be contained in the original ACL as
defined in VCL. Such log entries are marked by "fixed: merged".

implementation:

The sort functions are changed such that the previous semantics are
preserved: negative return values signify "a < b", positive return
values signify "a > b". But additional the values -2/2 and -3/3 are
introduced (and given enums) to signify "contained in supernet" and
"directly adjacent to". This allows for semantically mostly unchanged
code for vcc_acl_merge disabled.

For the "subnet contained in supernet" case, an acl entry is either
not added or existing entries removed.

For the "merge adjacent" case, the netmask of the entry with the
smaller network number is decreased by one and the other entry either
not added or removed.