Teach vcl its subs and allowed calling methods
This is in preparation for work to add a VCL_SUB vmod type which would
allow VMODs to call into vcl subs.

We add to VRT and initialize for each vcl sub a struct vcl_sub which
contains:

	- a methods bitmask defining which contexts this sub may be
	  called from
	- the name as seen from vcl
	- the function pointer to the vcl_func_f

The latter two are trivial.

About the methods bitmask:

It contains the VCL_MET_* bits of contexts which may potentially call
this sub and is intended for later use with checks like

	if (sub->methods & ctx->method) {
		sub->func(ctx);
		return;
	} else {
		VRT_fail(ctx, "not allowed here");
		return;
	}

For existing compile-time calls, we check if used objects and returned
actions are allowed within the respective calling context.

To build the methods bitmask, we begin with an all-one bitfield and
clear the bits of methods which would not be allowed for any used
object or returned action.