Improve VCL reference caching for labels
The previous caching code was written before labels existed: Caching a
VCL reference made sense only if it was for the active vcl.

We now also cache references to VCLs referenced by a label, which
implies them being warm and not discarded.

Note that, while this patch improves VCL reference caching when labels
are used, it remains opportunistic:

Without labels, the cached reference will basically always be used
unless a "vcl.use" is issued changing the active vcl.

But with labels, chances are lower that the worker will have the right
vcl reference cached. Yet, this mechanism is simple and should still
help reduce vcl_mtx contention.