mgt: Adding Linux-specific jailing
Adding the very initial version of linux specific jailing
based on seccomp. This is a WIP but at least varnish seems
to be running fine.

There's a few TODOs:
	* Need to lower the varnish and its friends privs
	  via setuid(2) similar to UNIX jail;
	* Need to split the seccomp filters in two pieces:
		1) For MGT;
		2) For CLD;
	* Need to make all compatible with automake;

The granularity of seccomp(2) filters are still not on what
I thik it's ideal, as both MGT and CLDs have the same filter.

Signed-off-by: Marco Benatto <mbenatto@redhat.com>