Change the VRT_call() recursion check to use a vbitmask
Previously, we used a PRIV_TASK for recursion detection. This method did
not require cooperation from the core vcl call code, but had the
disadvantages of being relatively expensive and breaking rollbacks.

We now a call bitmask in the vcl calling code if any SUB reference is
passed to a vmod. Thus, runtime recursion checking is only activated if
needed.