Give workspaces a generation count
One fundamental complication with use of workspaces (which we did get
wrong for quite some time in the past) is to consider the rollback case
and ensure that any objects on a workspace are finalized before rollback
and any pointers to objects on a workspace cleared.

Up to now, no way existed to localize handling of the rollback case in
the owner of an object, we neeed to hardcode it in Req_Rollback() and
Bereq_Rollback() and via the fini callback of VRT privs.

Adding a generation count allows us to localize rollback handling with
objects. Consider, for example, a data structure with a cache on a
workspace. We can now use the generation count to check if our cache is
still valid:

struct sth {
	struct ws	*ws;
	VCL_STRING	*cache;
	uint8_t		gen;
};

VCL_STRING
return_value(struct sth *sth)
{
	if (sth->cache != NULL &&
	    WS_Generation(sth->ws) == sth->gen)
		return (sth->cache);
	sth->cache = generate_value(sth);
	sth->gen = sth->ws->gen;
}