outlaw std.rollback(req) in vcl_pipe {}
With the previous commits, as the bereq uses the req workspace, we
cannot roll back the req without also rolling back the bereq.

As the use case is rather exotic, we spare the additional complications
and just outlaw rollbacks from vcl_pipe.

std.rollback(bereq) is already forbidden by the compiler because the
bereq symbol is not available in vcl_pipe {}.

We could, alternative to the runtime failure for std.rollback(req),
remove the req symbol from vcl_pipe {}, but that could break use of req
to vmod functions also. So we chose the more compatible route of
selectively failing rollback rather than making impossible other use
cases.