Add vcc_trace parameter to control generation of VCL_trace logs
To potentially emit VCL_trace VSL records, we called VPI_count() from
VGC for, in the extreme, every line of VCL.

For a normal setup, VPI_count() would call into VSL only to find out
that VCL_trace is masked and not have any effect.

Issuing two additional function calls for each line of VCL obviously
is something we should avoid unless there is a real benefit.

Under the assumption that the VCL_trace facility is used rarely, we
add the vcc_trace parameter which must explicitly be enabled before
loading a VCL to enable tracing. This should have no relevant impact
for common installations and, if tracing is required, re-loading the
vcl to enable tracing is quick and easy.

The benefit of this change is that we can turn VPI_count(), which
tracks an approximation of our position in VCL processing, into a
macro, reducing the overhead from the above mentioned two function
calls to a single address write (one of the two lines of VPI_count()
is an assertion on constant values which can (and usually will) be
eliminated by compilers).