Add std.now() and std.timed_call()
std.now() fills a gap left by the stable time of the now variable in
VCL: Sometimes we need to know the current time after some longer
processing in VCL.

std.timed_call() is intended to measure longer operations in
VCL. std.timestamp() exists already to add accurate measurements to
the log. std.timed_call() is for cases where the duration is required
in VCL, for example to enrich HTTP requests/responses.