Fix VRE_capture() interface wrt more matching groups than group count
The new VRE_capture() interface did not provide a way to detect the
case where the number of matching subexpressions was higher than the
group count passed.

This patch basically restores the original pcre (pcre1 if you will)
behaviour to return a number higher than the number of groups to
signal this case.

Alternatively, if we wanted to have the pcre2-like semantics (pcre2
returns 0 for this case), we would basically need to make the
VRE_capture() interface identical to vre_capture() and allow to pass a
count pointer, otherwise we had no way to return the number of
matches.