Explicitly test connection pooling works properly
Connect to s1 and s2 via v2.

Note on the v2 VCL: We use this varnish instance as a PROXY protocol
aware forwarder, which takes the address to connect to from the
incoming PROXY header (to mimic haproxy instead of requiring it).

Previously, we used debug.dyn(), but that does not work with two
different backends because it does not create different backend
instances, so connection pooling fails on this level, unrelated to the
actual test subject.

We avoid this issue by an explicit VCL implementation.