req.body: Don't cache an empty body
It's either guaranteed to fail a fetch or cache nothing, in the sense of
accessing to the request body through a VMOD. It is still possible to
end up with an empty body cached, if req.body turned out to be empty and
couldn't be predicted. It is no longer possible to actively try caching
an empty one.