Renovate autoconf
this is required for autoconf 2.71, but I have kept the required
version at 2.69 and tested that the changes are compatible.

Most of the changes come from autoupdate, with this change:

AC_PROG_CC_STDC is obsolete, but older autoconf versions
do not set the highest possible c standard.

So we keep AC_PROG_CC_C99 because we actually require
c99, while AC_PROG_CC_STDC could fall back to c89