Access to partially cached request body
This commit adds an argument to VRB_Iterate() if the full request body,
only the cached part or only the remaining part is to be iterated over.

This implements at least two relevant use cases:

1) The cached part of the request body can now be processed multiple
   times, as could a fully cached request body before introduction
   of partial caching.

2) VMODs can make an effort to consume only the cached part (to
   keep the request body intact for sending it to the backend),
   but can still opt to also process the remainder.

These cases are exemplified in vmod_debug.

In addition to the previously implemented cases, we can now also start
a backend request (as pipe or pass/fetch) with the request body already
taken, in which case we fail before sending the request.