hash: Trigger a rush when an objcore is withdrawn
This is the counterpart of HSH_Unbusy() for the cases where the req task
will not schedule a fetch task, at which point we know we can already
wake up a request, if there is one on the waiting list.

This encapsulates the "wake up only one request" semantic using a new
OC_F_WITHDRAWN flag. Although this flag is not being used yet, it will
when the state of the objcore determines the rush policy instead of the
call site.

When a bgfetch is withdrawn, the extra rush becomes redundant with the
policy rush on the hit objcore. There will eventually be no more rush
for the hit objcore, which will remove the redundancy just introduced.

In order to withdraw objcores in a single critical section, a function
HSH_DerefObjCoreUnlock() is added.