builtin.vcl: Delete Content-Range header where it has no meaning
Since 4ab110047130e3a89936104d74f4729b650676f9 (see also #3246), we fail backend
responses with an unexpected Content-Range.

Yet RFC9110 states:

	The Content-Range header field has no meaning for status codes that do
	not explicitly describe its semantic.

	https://www.rfc-editor.org/rfc/rfc9110.html#section-14.4

The semantic is described for status codes 206 and 416, so the obvious
implementation change would be for core code to only consider Content-Range for
these status codes.

But there might be scenarios where a stricter-than-RFC check is intended, so we
keep that in core code and change builtin.vcl to remove the header where it has
no semantic.