mgt: Move -a(cceptor) and -s(torage) configuration until after VEXT load
... to enable acceptor implementations as extensions (and also allow storage to
make use of manager privileges, but for this we probably need a new or
repurposed STV callback).

One test adjustment was required:

c3.vtc tests invalid -a usage, but did not set up a good working directory. Now
that -a is processed later, the test ran into a working directory error instead
of the expected acceptor error.