Refactor: make the V1L VDP a bit less special
V1L is used to send HTTP headers and the body, so it is used directly from
delivery/fetch first and then as a VDP.

From the times before VDPs, the V1L VDP still had its private pointer in struct
wrk.

This commit is to move the private pointer to the VDP entry and use the VDP
facilities for initialization and finalization. This assumes that any use of V1L
also implies use a VDP.