cache_http1_line: Make the v1l VDP a bit less special
V1L is used to send HTTP headers and the body, so it is used directly from
delivery/fetch first and then as a VDP.

From the times before VDPs, the V1L VDP still had its private pointer in struct
wrk.

This commit is to move the private pointer to the VDP entry. The caller remains
responsible for calling V1L_Close() to keep V1L independent of VDP.