cache_expire: add batching also for the expire case
Similar to the changes for cache removal batching, we now gather all object
cores to be expired in another head of the deref struct in one go within a
single critical region (as before, exp_list is free when the respective object
is not in the inbox) and run the derefs in the existing exp_deref_work()
function, which optionally gets offloaded to a separate worker thread.

For efficiency, we also change handling of expired object cores with
OC_EF_POSTED: Previously, they would just be ignored, only to be handled by the
exp_inbox(). But now more than ever it does not make sense to keep them around,
so we just remove them from the inbox and add them to our list to expire.

(diff best viewed ignoring whitespace (-b option))