cache_expire: Avoid double counting of OC_EF_REMOVE objects
When the exp_expire() finds the expired object also in the inbox, it would still
count it as expired. We now avoid the double count by excluding objects with
OC_EF_REMOVE set.

And we also reduce traffic on the global counter by using a local variable for
the loop.