cache_req{,_fsm}: Add resp.storage control over synth body's storage
This adds the vcl variable resp.storage, available from vcl_synth {} to control
which storage the synth response body gets created on.

Rather than just remembering the stroage for use after VCL has returned, we
actually create an object on it when the variable changes.

This is the first step towards #4344